requirements.txt
streamlitstreamlit
requests
beautifulsoup4
sentence-transformers
torch
scikit-learn